Submitter Registration List using React
