+++ apcd-cms/src/apps/admin_regis_table/templates/list_registrations.html
@@ -15,57 +15,4 @@
   <hr />
   <div id="list-registrations-root"></div>
 </div>
-<script>
-  function filterTableByStatus() {
-    var dropdown, statusFilter, xhr, url_params, url;
-    dropdown = document.getElementById("statusFilter");
-    statusFilter =  dropdown.value;
-    url_params = `?status=${statusFilter}`;
-    // {% if selected_org %}
-    //   url_params += `&org={{selected_org}}`;
-    // {% endif %}
-    url = `/administration/list-registration-requests/${url_params}`;
-    xhr = new XMLHttpRequest();
-    xhr.open('GET', url);
-    xhr.send();
-    window.location.href = url;
-    window.location.load();
-  }
-  function filterTableByOrganization() {
-    var input, orgFilter, xhr, url_params, url;
-    input = document.getElementById("organizationFilter");
-    orgFilter = input.value.replace("&", encodeURIComponent('&'));
-    url_params = `?org=${orgFilter}`;
-    // {% if selected_status %}
-    //   url_params = `?status={{selected_status}}&org=${orgFilter}`;
-    // {% endif %}
-    url = `/administration/list-registration-requests/${url_params}`;
-    xhr = new XMLHttpRequest();
-    xhr.open('GET', url);
-    xhr.send();
-    window.location.href = url;
-    window.location.load();
-  }
-  function clearSelections() {
-        var xhr;
-        xhr = new XMLHttpRequest();
-        xhr.open('GET', '/administration/list-registration-requests/')
-        xhr.send()
-        window.location.href = '/administration/list-registration-requests/';
-        window.location.load();
-    }
-
-  function openAction(reg_id) {
-    var actionsDropdown, selectedOption, modal_id;
-    actionsDropdown = document.getElementById(`actionsDropdown_${reg_id}`);
-    /* grabs dropdown option number selected by user via selectedIndex, 
-      then grabs actual value associated with that option via 
-      options[index].value */
-    selectedOption = actionsDropdown.options[actionsDropdown.selectedIndex].value;
-    modal_id = `${selectedOption}Modal_${reg_id}`;
-    $(`#${modal_id}`).modal({backdrop: "static"}); /* modal appears manually */
-    actionsDropdown.selectedIndex = 0; /* resets dropdown to display 'Select Action' again */
-  }
-</script>
-
 {% endblock %}
