+++ apcd-cms/src/apps/submitter_renewals_listing/templates/list_submitter_registrations.html
@@ -1,4 +1,4 @@
-{% extends "standard.html" %}
+{% extends "apcd_cms/templates/standard.html" %}
 {% load static %}
 
 {% block content %}
@@ -14,115 +14,6 @@
   <hr />
   <p style="margin-bottom: 30px">Current registration information on file for your organization is listed below. On the right of the screen, select the dropdown menu under “Actions” to view or renew your organization’s registration information.</p>
   <hr />
-  <div class="filter-container">
-    <div class="filter-content">
-      <span><b>Filter by Status: </b></span>
-      <select id="statusFilter" class="status-filter" onchange="filterTableByStatus()">
-        {% for option in status_options %}
-          <option class="dropdown-text" {% if option == selected_status %}selected{% endif %}>{{ option }}</option>
-        {% endfor %}
-      </select>
-      <span><b>Filter by Organization: </b></span>
-      <select id="organizationFilter" class="status-filter org-filter" onchange="filterTableByOrganization()">
-        {% for option in org_options %}
-          <option class="dropdown-text" {% if option == selected_org %}selected{% endif %}>{{ option }}</option>
-        {% endfor %}
-      </select>
-      {% if selected_status or selected_org %}
-        <button onclick="clearSelections()">Clear Options</button>
-      {% endif %}
-    </div>
-  </div>
-  <table id="submittersTable" class="registration-table submitters-table">
-      <thead>
-          <tr>
-              {% for k in header %}
-              <th>{{k}}</th>
-              {% endfor %}
-          </tr>
-      </thead>
-      <tbody>
-          {% for r in page %}
-              <tr>
-                  <td>{{r.biz_name}}</td>
-                  <td>{{r.year}}</td>
-                  <td>{{r.type}}</td>
-                  <td>{{r.location}}</td>
-                  <td>{{r.reg_status}}</td>
-                  <td>
-                    {% include "view_registration_modal.html" %}
-                    <select id='actionsSubmitterDropdown_{{r.reg_id}}' class='status-filter' onchange="openAction('{{r.reg_id}}')">
-                      <option value="">Actions</option>
-                      <option value="viewRegistration">View Record</option>
-                      <option value="renewRegistration">Renew Registration</option>
-                    </select>
-                  </td>
-              </tr>
-          {% endfor %}
-      </tbody>
-  </table>
-  {% include 'paginator.html' %}
+  <div id="list-submitter-registrations-root"></div>
 </div>
-<script>
-  function filterTableByStatus() {
-    var dropdown, statusFilter, xhr, url_params, url;
-    dropdown = document.getElementById("statusFilter");
-    statusFilter =  dropdown.value;
-    url_params = `?status=${statusFilter}`;
-    {% if selected_org %}
-      url_params += `&org={{selected_org}}`;
-    {% endif %}
-    url = `/register/list-registration-requests/${url_params}`;
-    xhr = new XMLHttpRequest();
-    xhr.open('GET', url);
-    xhr.send();
-    window.location.href = url;
-    window.location.load();
-  }
-  function filterTableByOrganization() {
-    var input, orgFilter, xhr, url_params, url;
-    input = document.getElementById("organizationFilter");
-    orgFilter = input.value.replace("&", encodeURIComponent('&'));
-    url_params = `?org=${orgFilter}`;
-    {% if selected_status %}
-      url_params = `?status={{selected_status}}&org=${orgFilter}`;
-    {% endif %}
-    url = `/register/list-registration-requests/${url_params}`;
-    xhr = new XMLHttpRequest();
-    xhr.open('GET', url);
-    xhr.send();
-    window.location.href = url;
-    window.location.load();
-  }
-  function clearSelections() {
-        var xhr;
-        xhr = new XMLHttpRequest();
-        xhr.open('GET', '/register/list-registration-requests/')
-        xhr.send()
-        window.location.href = '/register/list-registration-requests/';
-        window.location.load();
-    }
-
-    function openAction(reg_id) {
-      var actionsDropdown, selectedOption, modal_id;
-      actionsDropdown = document.getElementById(`actionsSubmitterDropdown_${reg_id}`);
-      /* grabs dropdown option number selected by user via selectedIndex, 
-        then grabs actual value associated with that option via 
-        options[index].value */
-      selectedOption = actionsDropdown.options[actionsDropdown.selectedIndex].value;
-      modal_id = `${selectedOption}Modal_${reg_id}`;
-      $(`#${modal_id}`).modal({backdrop: "static"}); /* modal appears manually */
-      actionsDropdown.selectedIndex = 0; /* resets dropdown to display 'Select Action' again */
-      if (selectedOption == "renewRegistration") {
-        var xhr, url;
-        url = `/register/request-to-submit/?reg_id=${reg_id}`
-        xhr = new XMLHttpRequest();
-        xhr.open('GET', url)
-        xhr.send()
-        window.location.href = url;
-        window.location.load();
-      }
-    }
-</script>
-
 {% endblock %}
