+++ apcd-cms/src/apps/submitter_renewals_listing/urls.py
@@ -1,10 +1,12 @@
 from django.urls import path
+from django.views.generic import TemplateView
 from apps.submitter_renewals_listing.views import SubmittersTable
 
 app_name = 'register'
 urlpatterns = [
-    path('list-registration-requests/', SubmittersTable.as_view(), name='submitter_regis_table'),
-    path(r'list-registration-requests/?status=(?P<status>)/', SubmittersTable.as_view(), name='submitter_regis_table'),
-    path(r'list-registration-requests/?org=(?P<org>)/', SubmittersTable.as_view(), name='submitter_regis_table'),
-    path(r'list-registration-requests/?status=(?P<status>)&org=(?P<org>)/', SubmittersTable.as_view(), name='submitter_regis_table')
+    path('list-registration-requests/', TemplateView.as_view(template_name='list_submitter_registrations.html'), name='submitter_regis_table'),
+    path(r'list-registration-requests/api/', SubmittersTable.as_view(), name='submitter_regis_table_api'),
+    path(r'list-registration-requests/api/?status=(?P<status>)/', SubmittersTable.as_view(), name='submitter_regis_table_api'),
+    path(r'list-registration-requests/api/?org=(?P<org>)/', SubmittersTable.as_view(), name='submitter_regis_table_api'),
+    path(r'list-registration-requests/api/?status=(?P<status>)&org=(?P<org>)/', SubmittersTable.as_view(), name='submitter_regis_table_api'),
 ]
