+++ apcd-cms/src/client/src/components/Admin/ViewUsers/EditRecordModal.tsx
@@ -1,10 +1,19 @@
 import React, { useState } from 'react';
-import { Modal, ModalBody, ModalFooter, Button, Label, FormGroup, Row, Col } from 'reactstrap';
+import {
+  Modal,
+  ModalBody,
+  ModalFooter,
+  Button,
+  Label,
+  FormGroup,
+  Row,
+  Col,
+} from 'reactstrap';
 import { Formik, Field, Form, ErrorMessage } from 'formik';
 import { fetchUtil } from 'utils/fetchUtil';
 import * as Yup from 'yup';
 import { UserRow } from 'hooks/admin';
-import styles from './ViewUsers.module.scss';  // Import SCSS module
+import styles from './ViewUsers.module.scss'; // Import SCSS module
 
 interface EditRecordModalProps {
   isOpen: boolean;
@@ -13,7 +22,12 @@ interface EditRecordModalProps {
   onEditSuccess?: (updatedUser: UserRow) => void;
 }
 
-const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user, onEditSuccess }) => {  
+const EditRecordModal: React.FC<EditRecordModalProps> = ({
+  isOpen,
+  toggle,
+  user,
+  onEditSuccess,
+}) => {
   const [successModalOpen, setSuccessModalOpen] = useState(false);
   const [errorModalOpen, setErrorModalOpen] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
@@ -22,18 +36,25 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
 
   const initialValues: UserRow = {
     ...user,
-    notes: user.notes || 'None',  // Set notes to 'None' if it is null
+    notes: user.notes || 'None', // Set notes to 'None' if it is null
   };
 
   const validationSchema = Yup.object({
     user_name: Yup.string().required('Name is required'),
-    user_email: Yup.string().email('Invalid email format').required('Email is required'),
+    user_email: Yup.string()
+      .email('Invalid email format')
+      .required('Email is required'),
     status: Yup.string().required('Status is required'),
     role_name: Yup.string().required('Role is required'),
-    notes: Yup.string().max(2000, 'Notes cannot exceed 2000 characters').nullable(),
+    notes: Yup.string()
+      .max(2000, 'Notes cannot exceed 2000 characters')
+      .nullable(),
   });
-  
-  const handleSave = async (values: UserRow, { setSubmitting }: { setSubmitting: (isSubmitting: boolean) => void }) => {
+
+  const handleSave = async (
+    values: UserRow,
+    { setSubmitting }: { setSubmitting: (isSubmitting: boolean) => void }
+  ) => {
     const { user_number } = values;
     const url = `administration/users/${user_number}/`;
     try {
@@ -42,28 +63,32 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
         method: 'PUT',
         body: values,
       });
-  
+
       if (onEditSuccess && response) {
-        onEditSuccess(response); 
+        onEditSuccess(response);
       }
-  
+
       setSuccessModalOpen(true);
     } catch (error: any) {
       console.error('Error saving data:', error);
       console.log(url);
       if (error.response && error.response.data) {
         // Use error message from the server response
-        setErrorMessage(error.response.data.message || 'An error occurred while saving the data. Please try again.');
+        setErrorMessage(
+          error.response.data.message ||
+            'An error occurred while saving the data. Please try again.'
+        );
       } else {
         // Use generic error message
-        setErrorMessage('An error occurred while saving the data. Please try again.');
+        setErrorMessage(
+          'An error occurred while saving the data. Please try again.'
+        );
       }
       setErrorModalOpen(true);
     } finally {
       setSubmitting(false);
     }
   };
-  
 
   const userFields = [
     { label: 'User ID', value: user.user_id },
@@ -75,7 +100,7 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
     { label: 'Status', value: user.status },
     { label: 'Created Date', value: user.created_at },
     { label: 'Updated Date', value: user.updated_at },
-    { label: 'Notes', value: user.notes ? user.notes : "None" },
+    { label: 'Notes', value: user.notes ? user.notes : 'None' },
   ];
 
   return (
@@ -85,14 +110,16 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
           <Label className={styles.customModalTitle}>
             Edit User ID: {user.user_id} for {user.user_name}
           </Label>
-          <button type="button" className={`close ${styles.customCloseButton}`} onClick={toggle}>
+          <button
+            type="button"
+            className={`close ${styles.customCloseButton}`}
+            onClick={toggle}
+          >
             <span aria-hidden="true">&times;</span>
           </button>
         </div>
         <ModalBody>
-          <div className={styles.greyRectangle}>
-            Edit Selected User
-          </div>
+          <div className={styles.greyRectangle}>Edit Selected User</div>
           <Formik
             initialValues={initialValues}
             validationSchema={validationSchema}
@@ -103,31 +130,45 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
                 <Row>
                   <Col md={3}>
                     <FormGroup>
-                      <Label for="user_name" className={styles.customLabel}><strong>Name</strong></Label>
+                      <Label for="user_name" className={styles.customLabel}>
+                        <strong>Name</strong>
+                      </Label>
                       <Field
                         type="text"
                         name="user_name"
                         id="user_name"
                         className={`form-control ${styles.viewRecord}`}
                       />
-                      <ErrorMessage name="user_name" component="div" className="text-danger" />
+                      <ErrorMessage
+                        name="user_name"
+                        component="div"
+                        className="text-danger"
+                      />
                     </FormGroup>
                   </Col>
                   <Col md={3}>
                     <FormGroup>
-                      <Label for="user_email" className={styles.customLabel}><strong>Email</strong></Label>
+                      <Label for="user_email" className={styles.customLabel}>
+                        <strong>Email</strong>
+                      </Label>
                       <Field
                         type="email"
                         name="user_email"
                         id="user_email"
                         className={`form-control ${styles.viewRecord}`}
                       />
-                      <ErrorMessage name="user_email" component="div" className="text-danger" />
+                      <ErrorMessage
+                        name="user_email"
+                        component="div"
+                        className="text-danger"
+                      />
                     </FormGroup>
                   </Col>
                   <Col md={2}>
                     <FormGroup>
-                      <Label for="status" className={styles.customLabel}><strong>Active Status</strong></Label>
+                      <Label for="status" className={styles.customLabel}>
+                        <strong>Active Status</strong>
+                      </Label>
                       <Field
                         as="select"
                         name="status"
@@ -137,12 +178,18 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
                         <option value="Active">Active</option>
                         <option value="Inactive">Inactive</option>
                       </Field>
-                      <ErrorMessage name="status" component="div" className="text-danger" />
+                      <ErrorMessage
+                        name="status"
+                        component="div"
+                        className="text-danger"
+                      />
                     </FormGroup>
                   </Col>
                   <Col md={3}>
                     <FormGroup>
-                      <Label for="role_name" className={styles.customLabel}><strong>Role</strong></Label>
+                      <Label for="role_name" className={styles.customLabel}>
+                        <strong>Role</strong>
+                      </Label>
                       <Field
                         as="select"
                         name="role_name"
@@ -153,31 +200,49 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
                         <option value="SUBMITTER_ADMIN">SUBMITTER_ADMIN</option>
                         <option value="APCD_ADMIN">APCD_ADMIN</option>
                       </Field>
-                      <ErrorMessage name="role_name" component="div" className="text-danger" />
+                      <ErrorMessage
+                        name="role_name"
+                        component="div"
+                        className="text-danger"
+                      />
                     </FormGroup>
                   </Col>
                 </Row>
                 <Row>
                   <Col md={6}>
                     <FormGroup>
-                      <Label for="notes" className={styles.customLabel}><strong>Notes</strong></Label>
+                      <Label for="notes" className={styles.customLabel}>
+                        <strong>Notes</strong>
+                      </Label>
                       <Field
                         as="textarea"
                         name="notes"
                         id="notes"
                         rows="5"
-                        maxLength="2000"  // Set the maxLength attribute
+                        maxLength="2000" // Set the maxLength attribute
                         className={`form-control ${styles.viewRecord}`}
                       />
-                      <small className="form-text text-muted" style={{ fontStyle: 'italic' }}>
+                      <small
+                        className="form-text text-muted"
+                        style={{ fontStyle: 'italic' }}
+                      >
                         2000 character limit
                       </small>
-                      <ErrorMessage name="notes" component="div" className="text-danger" />
+                      <ErrorMessage
+                        name="notes"
+                        component="div"
+                        className="text-danger"
+                      />
                     </FormGroup>
                   </Col>
                 </Row>
                 <br />
-                <Button type="submit" color="primary" disabled={isSubmitting} className={styles.customSubmitButton}>
+                <Button
+                  type="submit"
+                  color="primary"
+                  disabled={isSubmitting}
+                  className={styles.customSubmitButton}
+                >
                   Submit
                 </Button>
               </Form>
@@ -199,32 +264,40 @@ const EditRecordModal: React.FC<EditRecordModalProps> = ({ isOpen, toggle, user,
         </ModalBody>
       </Modal>
 
-      <Modal isOpen={successModalOpen} toggle={() => setSuccessModalOpen(false)} className={styles.customModal}>
+      <Modal
+        isOpen={successModalOpen}
+        toggle={() => setSuccessModalOpen(false)}
+        className={styles.customModal}
+      >
         <div className={`modal-header ${styles.modalHeader}`}>
-            <Label  className={styles.customModalTitle}>
-                Success
-            </Label>
-            <button type="button" className={`close ${styles.customCloseButton}`} onClick={toggle}>
-                <span aria-hidden="true">&#xe912;</span>
-            </button>
+          <Label className={styles.customModalTitle}>Success</Label>
+          <button
+            type="button"
+            className={`close ${styles.customCloseButton}`}
+            onClick={toggle}
+          >
+            <span aria-hidden="true">&#xe912;</span>
+          </button>
         </div>
-        <ModalBody>
-          The user data has been successfully updated.
-        </ModalBody>
+        <ModalBody>The user data has been successfully updated.</ModalBody>
       </Modal>
 
-      <Modal isOpen={errorModalOpen} toggle={() => setErrorModalOpen(false)} className={styles.customModal}>
+      <Modal
+        isOpen={errorModalOpen}
+        toggle={() => setErrorModalOpen(false)}
+        className={styles.customModal}
+      >
         <div className={`modal-header ${styles.modalHeader}`}>
-            <Label className={styles.customModalTitle}>
-                Error
-            </Label>
-            <button type="button" className={`close ${styles.customCloseButton}`} onClick={toggle}>
-                <span aria-hidden="true">&#xe912;</span>
-            </button>
+          <Label className={styles.customModalTitle}>Error</Label>
+          <button
+            type="button"
+            className={`close ${styles.customCloseButton}`}
+            onClick={toggle}
+          >
+            <span aria-hidden="true">&#xe912;</span>
+          </button>
         </div>
-        <ModalBody>
-          {errorMessage}
-        </ModalBody>
+        <ModalBody>{errorMessage}</ModalBody>
       </Modal>
     </>
   );
