+++ apcd-cms/src/client/src/components/Admin/ViewUsers/ViewRecordModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Modal, ModalHeader, ModalBody, Label } from 'reactstrap';
 import { UserRow } from 'hooks/admin';
-import styles from './ViewUsers.module.scss';  // Import SCSS module
+import styles from './ViewUsers.module.scss'; // Import SCSS module
 
 interface UserDetailsModalProps {
   isOpen: boolean;
@@ -9,7 +9,11 @@ interface UserDetailsModalProps {
   user: UserRow | null;
 }
 
-const UserDetailsModal: React.FC<UserDetailsModalProps> = ({ isOpen, toggle, user }) => {
+const UserDetailsModal: React.FC<UserDetailsModalProps> = ({
+  isOpen,
+  toggle,
+  user,
+}) => {
   if (!user) return null;
 
   return (
@@ -18,21 +22,45 @@ const UserDetailsModal: React.FC<UserDetailsModalProps> = ({ isOpen, toggle, use
         <Label className={styles.customModalTitle}>
           Details for User: {user.user_name} ({user.user_id})
         </Label>
-        <button type="button" className={`close ${styles.customCloseButton}`} onClick={toggle}>
+        <button
+          type="button"
+          className={`close ${styles.customCloseButton}`}
+          onClick={toggle}
+        >
           <span aria-hidden="true">&#xe912;</span>
         </button>
       </div>
       <ModalBody>
-        <p><strong>User ID:</strong> {user.user_id}</p>
-        <p><strong>Name:</strong> {user.user_name}</p>
-        <p><strong>User Number:</strong> {user.user_number}</p>
-        <p><strong>Email:</strong> {user.user_email}</p>
-        <p><strong>Entity Organization:</strong> {user.entity_name}</p>
-        <p><strong>Role:</strong> {user.role_name}</p>
-        <p><strong>Status:</strong> {user.status}</p>
-        <p><strong>Created Date:</strong> {user.created_at}</p>
-        <p><strong>Updated Date:</strong> {user.updated_at}</p>
-        <p><strong>Notes:</strong> {user.notes ? user.notes : "None"}</p>
+        <p>
+          <strong>User ID:</strong> {user.user_id}
+        </p>
+        <p>
+          <strong>Name:</strong> {user.user_name}
+        </p>
+        <p>
+          <strong>User Number:</strong> {user.user_number}
+        </p>
+        <p>
+          <strong>Email:</strong> {user.user_email}
+        </p>
+        <p>
+          <strong>Entity Organization:</strong> {user.entity_name}
+        </p>
+        <p>
+          <strong>Role:</strong> {user.role_name}
+        </p>
+        <p>
+          <strong>Status:</strong> {user.status}
+        </p>
+        <p>
+          <strong>Created Date:</strong> {user.created_at}
+        </p>
+        <p>
+          <strong>Updated Date:</strong> {user.updated_at}
+        </p>
+        <p>
+          <strong>Notes:</strong> {user.notes ? user.notes : 'None'}
+        </p>
       </ModalBody>
     </Modal>
   );
