+++ apcd-cms/src/client/src/components/Admin/ViewUsers/ViewUsers.module.scss
@@ -12,18 +12,18 @@ $modal-width: 55%;
 }
 
 .customModalTitle {
-  text-transform: none !important;  // Prevent text from being all caps
+  text-transform: none !important; // Prevent text from being all caps
   font-size: 2.5rem;
-  margin: 0;  // Ensure there's no margin
+  margin: 0; // Ensure there's no margin
 }
 
 .customLabel {
-  font-size: 1.5rem;  // Adjust the font size as needed
+  font-size: 1.5rem; // Adjust the font size as needed
 }
 
 .customSubmitButton {
-  font-size: 1.5rem;  // Increase font size
-  padding: 0.75rem 1.5rem;  // Increase padding
+  font-size: 1.5rem; // Increase font size
+  padding: 0.75rem 1.5rem; // Increase padding
 }
 
 .viewRecord {
