+++ apcd-cms/src/client/src/components/Registrations/RegistrationList/index.ts
@@ -1,0 +1,4 @@
+// index.ts
+import { RegistrationList } from './RegistrationList';
+
+export { RegistrationList };
