+++ apcd-cms/src/client/src/components/Submitter/Registrations/SubmitterRegistrationList.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { RegistrationList } from 'apcd-components/Registrations/RegistrationList';
+import { useSubmitterRegistrations } from 'hooks/registrations';
+
+export const SubmitterRegistrationList: React.FC = () => {
+  return <RegistrationList useDataHook={useSubmitterRegistrations} />;
+};
