+++ apcd-cms/src/client/src/components/Submitter/Registrations/index.ts
@@ -1,0 +1,4 @@
+// index.ts
+import { SubmitterRegistrationList } from './SubmitterRegistrationList';
+
+export { SubmitterRegistrationList };
