+++ apcd-cms/src/client/src/hooks/admin/useAdmin.ts
@@ -2,34 +2,10 @@ import { useQuery, UseQueryResult } from 'react-query';
 import { fetchUtil } from 'utils/fetchUtil';
 import {
   ExtensionResult,
-  RegistrationResult,
   UserResult,
   SubmissionResult,
   ExceptionResult,
 } from '.';
-
-const getRegistrations = async (params: any) => {
-  const url = `/administration/list-registration-requests/api/`;
-  const response = await fetchUtil({ url, params });
-  return response.response;
-};
-
-export const useRegistrations = (
-  status?: string,
-  org?: string,
-  page?: number
-): UseQueryResult<RegistrationResult> => {
-  const params: { status?: string; org?: string; page?: number } = {
-    status,
-    org,
-    page,
-  };
-  const query = useQuery(['registrations', params], () =>
-    getRegistrations(params)
-  ) as UseQueryResult<RegistrationResult>;
-
-  return { ...query };
-};
 
 const getUsers = async (params: any) => {
   const url = `/administration/view-users/api/`;
@@ -76,9 +52,7 @@ const getExtensions = async (params: any) => {
   return response.response;
 };
 
-export const useExtensions = (
-  params: any
-): UseQueryResult<ExtensionResult> => {
+export const useExtensions = (params: any): UseQueryResult<ExtensionResult> => {
   const query = useQuery(['extensions', params], () =>
     getExtensions(params)
   ) as UseQueryResult<ExtensionResult>;
