+++ apcd-cms/src/client/src/main.tsx
@@ -7,6 +7,7 @@ import { ViewUsers } from './components/Admin/ViewUsers';
 import { AdminSubmissions } from './components/Admin/Submissions';
 import { AdminExtensions } from './components/Admin/Extensions';
 import { AdminExceptions } from './components/Admin/Exceptions';
+import { SubmitterRegistrationList } from './components/Submitter/Registrations';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
 const queryClient = new QueryClient();
@@ -32,6 +33,7 @@ const componentMap: { [key: string]: React.ComponentType<any> } = {
   'list-admin-submissions': AdminSubmissions,
   'admin-extensions-root': AdminExtensions,
   'admin-exceptions-root': AdminExceptions,
+  'list-submitter-registrations-root': SubmitterRegistrationList,
   // Add new components with html id in the list above.
 };
